[IMP] round

+++ models/truck.py
@@ -9,8 +9,8 @@ class Truck(models.AbstractModel):
     car_plates = fields.Char()
     date = fields.Date(default=fields.Date.today)
 
-    input_kilos = fields.Float()
-    output_kilos = fields.Float()
+    input_kilos = fields.Integer()
+    output_kilos = fields.Integer()
 
     @api.one
     @api.depends('input_kilos', 'output_kilos')
